+++ src/components/EChartsEditor2/index.ts
