+++ src/components/EChartsPanel2/EChartsPanel.test.tsx
