+++ src/components/EChartsPanel2/EChartsPanel.tsx
