+++ src/components/EChartsPanel2/index.ts
