+++ src/components/index.ts
@@ -1,2 +1,2 @@
-export * from './EChartsEditor';
-export * from './EChartsPanel';
+export * from './EChartsEditor2';
+export * from './EChartsPanel2';
